requirements.txt
Django==2.2.6
django-rest-swagger==2.2.0
djangorestframework==3.10.3
pymongo==3.9.0
